refactor(dashboard): decompose server into feed-engine, http-router and socket-handlers (A4)
